requirements.txt
python-dotenv==0.20.0
uvicorn[standard]==0.17.6
fastapi==0.78.0
SQLAlchemy==1.4.36
python-multipart==0.0.5
minio==7.1.8
pillow==9.1.1
requests==2.27.1
pytest==7.1.2 